+++ Runtime/Scripts/GltFast.cs
@@ -82,9 +82,6 @@ namespace GLTFast {
         public const string ErrorUnsupportedColorFormat = "Unsupported Color format {0}";
         const string ErrorUnsupportedPrimitiveMode = "Primitive mode {0} is untested!";
         const string ErrorMissingImageURL = "Image URL missing";
-#if !KTX_UNITY
-        const string ErrorKtxUnsupported = "KTX textures are not supported!";
-#endif
         const string ErrorPackageMissing = "{0} package needs to be installed in order to support glTF extension {1}!\nSee https://github.com/atteneder/glTFast#installing for instructions";
 
         const string ExtDracoMeshCompression = "KHR_draco_mesh_compression";
@@ -254,7 +251,7 @@ namespace GLTFast {
                 }
 
                 if (gltfBinary ?? false) {
-                    success = await LoadGltfBinary(download.data,url);
+                    success = await LoadGltfBinaryBuffer(download.data,url);
                 } else {
                     success = await LoadGltf(download.text,url);
                 }
@@ -501,7 +498,19 @@ namespace GLTFast {
                 foreach(var ext in gltfRoot.extensionsUsed) {
                     var supported = supportedExtensions.Contains(ext);
                     if(!supported) {
-                        Debug.LogWarningFormat("glTF extension {0} is not supported!",ext);
+#if !DRACO_UNITY
+                        if(ext==ExtDracoMeshCompression) {
+                            Debug.LogWarningFormat(ErrorPackageMissing,"DracoUnity",ext);
+                        } else
+#endif
+#if !KTX_UNITY
+                        if(ext==ExtTextureBasisu) {
+                            Debug.LogWarningFormat(ErrorPackageMissing,"KtxUnity",ext);
+                        } else
+#endif
+                        {
+                            Debug.LogWarningFormat("glTF extension {0} is not supported!",ext);
+                        }
                     }
                 }
             }
@@ -835,7 +844,7 @@ namespace GLTFast {
                 }
                 ktxDownloadTasks.Add(index, downloadTask);
 #else
-                Debug.LogError(ErrorKtxUnsupported);
+                Debug.LogErrorFormat(ErrorPackageMissing,"KtxUnity",ExtTextureBasisu);
                 Profiler.EndSample();
                 return;
 #endif // KTX_UNITY
@@ -853,9 +862,19 @@ namespace GLTFast {
         /// Load a glTF-binary asset from a byte array.
         /// </summary>
         /// <param name="bytes">byte array containing glTF-binary</param>
-        /// <param name="url">Base URL for relative paths of external buffers or images</param>
-        /// <returns></returns>
-        public async Task<bool> LoadGltfBinary( byte[] bytes, Uri uri = null ) {
+        /// <param name="uri">Base URI for relative paths of external buffers or images</param>
+        /// <returns>True if loading was successful, false otherwise</returns>
+        public async Task<bool> LoadGltfBinary(byte[] bytes, Uri uri = null) {
+            var success = await LoadGltfBinaryBuffer(bytes,uri);
+            if(success) await LoadContent();
+            success = success && await Prepare();
+            DisposeVolatileData();
+            loadingError = !success;
+            loadingDone = true;
+            return success;
+        }
+        
+        async Task<bool> LoadGltfBinaryBuffer( byte[] bytes, Uri uri = null ) {
             Profiler.BeginSample("LoadGltfBinary.Phase1");
             uint magic = BitConverter.ToUInt32( bytes, 0 );
 
@@ -1323,7 +1342,7 @@ namespace GLTFast {
                             Profiler.EndSample();
                             await deferAgent.BreakPoint();
 #else
-                            Debug.LogError(ErrorKtxUnsupported);
+                            Debug.LogErrorFormat(ErrorPackageMissing,"KtxUnity",ExtTextureBasisu);
 #endif // KTX_UNITY
                         } else {
                             Profiler.BeginSample("CreateTexturesFromBuffers.ExtractBuffer");
@@ -1719,7 +1738,7 @@ namespace GLTFast {
                     i++;
                 }
             }
-            // TODO: not necesary with ECS
+            // TODO: not necessary with ECS
             // https://docs.unity3d.com/Manual/JobSystemTroubleshooting.html
             if (schedule) {
                 JobHandle.ScheduleBatchedJobs();