+++ Runtime/Scripts/PrimitiveCreateContext.cs
@@ -53,11 +53,11 @@ namespace GLTFast {
             var msh = new UnityEngine.Mesh();
             msh.name = mesh.name;
 
-            MeshUpdateFlags flags = MeshUpdateFlags.DontNotifyMeshUsers | MeshUpdateFlags.DontRecalculateBounds | MeshUpdateFlags.DontResetBoneBounds | MeshUpdateFlags.DontValidateIndices;
-            vertexData.ApplyOnMesh(msh,flags);
+            vertexData.ApplyOnMesh(msh,defaultMeshUpdateFlags);
 
             Profiler.BeginSample("SetIndices");
             int indexCount = 0;
+            var allBounds = vertexData.bounds;
             for (int i = 0; i < indices.Length; i++) {
                 indexCount += indices[i].Length;
             }
@@ -68,10 +68,26 @@ namespace GLTFast {
             indexCount = 0;
             for (int i = 0; i < indices.Length; i++) {
                 Profiler.BeginSample("SetIndexBufferData");
-                msh.SetIndexBufferData(indices[i],0,indexCount,indices[i].Length,flags);
+                msh.SetIndexBufferData(indices[i],0,indexCount,indices[i].Length,defaultMeshUpdateFlags);
                 Profiler.EndSample();
                 Profiler.BeginSample("SetSubMesh");
-                msh.SetSubMesh(i,new SubMeshDescriptor(indexCount,indices[i].Length,topology),flags);
+                var subMeshDescriptor = new SubMeshDescriptor{
+                    indexStart = indexCount,
+                    indexCount = indices[i].Length,
+                    topology = topology,
+                    baseVertex = 0,
+                    firstVertex = 0,
+                    vertexCount = vertexData.vertexCount
+                };
+                if (allBounds.HasValue) {
+                    // Setting the submeshes' bounds to the overall bounds
+                    // Calculating the actual sub-mesh bounds (by iterating the verts referenced
+                    // by the sub-mesh indices) would be slow. Also, hardly any glTFs re-use
+                    // the same vertex buffer across primitives of a node (which is the
+                    // only way a mesh can have sub-meshes)
+                    subMeshDescriptor.bounds = allBounds.Value;
+                }
+                msh.SetSubMesh(i,subMeshDescriptor,defaultMeshUpdateFlags);
                 Profiler.EndSample();
                 indexCount += indices[i].Length;
             }
@@ -88,9 +104,16 @@ namespace GLTFast {
                 Profiler.EndSample();
             }
             
-            Profiler.BeginSample("RecalculateBounds");
-            msh.RecalculateBounds(); // TODO: make optional! maybe calculate bounds in Job.
-            Profiler.EndSample();
+            if (allBounds.HasValue) {
+                msh.bounds = allBounds.Value;
+            } else {
+                Profiler.BeginSample("RecalculateBounds");
+#if DEBUG
+                Debug.LogError("Bounds have to be recalculated (slow operation). Check if position accessors have proper min/max values");
+#endif
+                msh.RecalculateBounds();
+                Profiler.EndSample();
+            }
 
 #if GLTFAST_KEEP_MESH_DATA
             Profiler.BeginSample("UploadMeshData");