+++ ChangeLog.md
@@ -4,13 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 
+## [3.1.0] - 2020-03-16
 ### Added
 - Unlit alpha blended ShaderGraph variants (thanks [@hybridherbst](https://github.com/hybridherbst) for #144)
+- Support for unsigned byte joint indices
+### Changed
+- Accelerated loading meshes by obtaining and setting bounds from accessors min/max values instead of recalculating them  
+- Improved log message when DracoUnity/KtxUnity packages are missing
+- Restored/simplified `GLTFast.LoadGltfBinary`, allowing users to load glTF binary files from byte arrays directly (also added documentation; fixes #148)
 ### Fixed
 - Texture offset/tiling values don't get lost when switching shaders (thanks [@hybridherbst](https://github.com/hybridherbst) for #140)
 - Correct vertex colors for RGB/unsigned short, RGBA/unsigned short and RGBA/unsigned byte. (thanks [@camogram](https://github.com/camogram) for #139)
-- Error when tryting to set texture offset/scale but material doesn't have _MainTex property (thanks [@hybridherbst](https://github.com/hybridherbst) for #142)
+- Error when trying to set texture offset/scale but material doesn't have _MainTex property (thanks [@hybridherbst](https://github.com/hybridherbst) for #142)
+- Crash when trying to combine meshes created by glTFast by setting proper submesh vertex count (fixes #100)
 
 ## [3.0.2] - 2020-02-07
 ### Changed